C functions 8

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,17 +7,17 @@
 int main(void)
 {
 	int count;
-	unsigned long long int a = 0;
-	unsigned long long int b = 1;
-	unsigned long long int c;
+	unsigned long int a = 0;
+	unsigned long int b = 1;
+	unsigned long int c;
 
-	printf("%llu, %llu", a, b);
+	printf("%lu, %lu", a, b);
 	for (count = 3; count <= 50; count++)
 	{
 		c = a + b;
 		a = b;
 		b = c;
-		printf(", %llu", c);
+		printf(", %lu", c);
 	}
 	printf("\n");
 	return (0);
